リファクタリング

+++ src/JpDrvLicNfcCommand/jpdlc_base_reader_if.h
@@ -8,8 +8,8 @@
 //カードリーダーをインクルード
 #include "rcs660s_app_if.h"
 
-#define READER_MAX_COMMAND_LEN 250
-#define READER_MAX_RES_LEN 250
+#define READER_MAX_COMMAND_LEN (uint16_t)250
+#define READER_MAX_RES_LEN     (uint16_t)250
 
 void setReaderInstance(Rcs660sAppIf *);
 std::vector<type_data_byte> _nfcTransceive(const std::vector<type_data_byte> );
